+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.35
 ### is.codion.common.core
 - Value.Notify.WHEN_CHANGED and WHEN_SET renamed CHANGED and SET respectively.
+- MethodLogger.Entry.getChildren() renamed children(), enterMessage() renamed message(), ArgumentToString renamed ArgumentFormatter.
 ### is.codion.common.db
 - Database.URL_SCOPED_INSTANCE configuration value added.
 ### is.codion.swing.common.model